refactor: extract reusable novel engine
